enable/disable without refreshing page + add menu button + restructure (#58)
* remove need to restart when (dis)enabling

Along the way restructure the package so that
codemirror keybindings are kept separated from
jlab command mode bindings.

* keep track of lastactive cell to switch it

Otherwise that one cell wouldn't be changed when you
modified the settings.

* add menu button for enable/diable

* Move keybindings into settings

Now users can override them!

* namespace the commands we add

Now is the one chance we get before people start using these.

* add docs for modifying keybindings

* move special thanks down

* refresh install instructions

* proofread

* update history
