+++ installing.md
@@ -1,0 +1,26 @@
+# Installing
+## For jupyterlab 3+
+
+```bash
+pip install jupyterlab_vim
+```
+
+Uninstall:
+
+```bash
+pip uninstall jupyterlab-vim
+```
+
+
+## For jupyterlab 2
+
+```bash
+jupyter labextension install @axlair/jupyterlab_vim
+```
+
+Uninstall:
+
+```bash
+jupyter labextension uninstall @axlair/jupyterlab_vim
+```
+
